Moved TransferAllowance->cfg_traits, fixed toml deps
