+++ pallets/transfer-allowlist/Cargo.toml
@@ -18,23 +18,20 @@ scale-info = { version = "2.3.0", default-features = false, features = ["derive"
 serde = { version = "1.0.119" }
 
 cfg-traits = { path = "../../libs/traits", default-features = false }
-pallet-connectors = { path = "../connectors" }
+
+cfg-types = { path = "../../libs/types", default-features = false }
 
 frame-benchmarking = { git = "https://github.com/paritytech/substrate", default-features = false, optional = true, branch = "polkadot-v0.9.32" }
 frame-support = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.32" }
 frame-system = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.32" }
+pallet-balances = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.32" }
 sp-core = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.32" }
 sp-runtime = { git = "https://github.com/paritytech/substrate", default-features = false, optional = true, branch = "polkadot-v0.9.32" }
 sp-std = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.32" }
 xcm = { git = "https://github.com/paritytech/polkadot", default-features = false, branch = "release-v0.9.32" }
 
-# Benchmarking dependencies
-cfg-primitives = { path = "../../libs/primitives", default-features = false, optional = false }
-cfg-types = { path = "../../libs/types", default-features = false, optional = true }
-
 [dev-dependencies]
-cfg-types = { path = "../../libs/types", default-features = true }
-pallet-balances = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.32" }
+pallet-balances = { git = "https://github.com/paritytech/substrate", default-features = true, branch = "polkadot-v0.9.32" }
 sp-io = { git = "https://github.com/paritytech/substrate", default-features = true, branch = "polkadot-v0.9.32" }
 
 [features]
@@ -42,7 +39,8 @@ default = ['std']
 runtime-benchmarks = [
   'frame-benchmarking',
   'sp-runtime',
-  'cfg-types',
+  'cfg-types/runtime-benchmarks',
+  'cfg-traits/runtime-benchmarks',
 ]
 std = [
   'codec/std',
@@ -55,4 +53,9 @@ std = [
   'cfg-traits/std',
 ]
 
-try-runtime = ['frame-support/try-runtime']
+try-runtime = [
+  'frame-support/try-runtime',
+  'frame-support/try-runtime',
+  'cfg-types/try-runtime',
+  'cfg-traits/try-runtime',
+]
